Fixed Tables, written some code for SelectFileWindow

+++ Creating XML/Creating XML/src/db/tables/CategoryParametersTable.cs
@@ -5,10 +5,10 @@ namespace Creating_XML.src.db.tables
     class CategoryParametersTable : Table
     {
         [AutoIncrement, PrimaryKey]
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         [Indexed, NotNull]
-        public long CategoryId { get; set; }
+        public int CategoryId { get; set; }
         
         [NotNull]
         public string Name { get; set; }
