+++ Creating XML/Creating XML/src/db/tables/CategoryTable.cs
@@ -5,10 +5,10 @@ namespace Creating_XML.src.db.tables
     class CategoryTable : Table
     {
         [AutoIncrement, PrimaryKey]
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         [Indexed]
-        public long ParentId { get; set; }
+        public int ParentId { get; set; }
 
         [Unique, NotNull]
         public string Name { get; set; }
